WIP add partition test
