--- src/test/scala/com/lucidworks/spark/CloudFrameworkSuite.scala
@@ -1,6 +1,0 @@
-package com.lucidworks.spark
-
-
-trait CloudFrameworkSuite extends {
-
-}
