+++ src/test/scala/com/lucidworks/spark/RDDTestSuite.scala
@@ -1,16 +1,37 @@
 package com.lucidworks.spark
 
+import org.apache.spark.Logging
 
-class RDDTestSuite extends SparkSolrFunSuite with SparkSolrContextBuilder {
+class RDDTestSuite extends SparkSolrFunSuite with SparkSolrContextBuilder with Logging {
 
-  test("Test RDD") {
-    val zkHost = cluster.getZkServer.getZkAddress
-    val collectioName = "test_new_rdd"
-    deleteCollection(collectioName)
-    buildCollection(zkHost, collectioName)
-    val newRDD = new SolrScalaRDD(zkHost, collectioName, sc)
-                                .query("*:*")
-    assert(newRDD.count() === 3)
-    deleteCollection(collectioName)
+  var zkHost: String = _
+
+  override def beforeAll(): Unit = {
+    super.beforeAll()
+    zkHost = cluster.getZkServer.getZkAddress
   }
+
+  test("Test Simple Query") {
+    val collectionName = "testSimpleQuery"
+    buildCollection(zkHost, collectionName, 3, 2)
+    try {
+      val newRDD = new SolrScalaRDD(zkHost, collectionName, sc)
+      assert(newRDD.count() === 3)
+    } finally {
+      deleteCollection(collectionName)
+    }
+  }
+
+  test("Test RDD Partitions") {
+    val collectionName = "testRDDPartitions"
+    buildCollection(zkHost, collectionName, 2, 4)
+    try {
+      val newRDD = new SolrScalaRDD(zkHost, collectionName, sc)
+      val partitions = newRDD.partitions
+      assert(partitions.length === 4)
+    } finally {
+      deleteCollection(collectionName)
+    }
+  }
+
 }
